+++ src/Factory/FormFactory.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Factory;
 
-use Innmind\Http\Message\Form;
+use Innmind\Http\ServerRequest\Form;
 
 /**
  * @psalm-immutable