+++ tests/ServerRequest/QueryTest.php
@@ -1,39 +1,39 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\Http\Message;
+namespace Tests\Innmind\Http\ServerRequest;
 
-use Innmind\Http\Message\Form;
+use Innmind\Http\ServerRequest\Query;
 use PHPUnit\Framework\TestCase;
 
-class FormTest extends TestCase
+class QueryTest extends TestCase
 {
     public function testInterface()
     {
-        $form = Form::of([
+        $query = Query::of([
             42 => '24',
             'foo' => 'bar',
         ]);
 
-        $this->assertFalse($form->contains('24'));
-        $this->assertTrue($form->contains(42));
-        $this->assertSame('24', $form->get(42)->match(
+        $this->assertFalse($query->contains('24'));
+        $this->assertTrue($query->contains(42));
+        $this->assertSame('24', $query->get(42)->match(
             static fn($value) => $value,
             static fn() => null,
         ));
-        $this->assertSame(2, $form->count());
+        $this->assertSame(2, $query->count());
         $this->assertSame(
             [
                 42 => '24',
                 'foo' => 'bar',
             ],
-            $form->data(),
+            $query->data(),
         );
     }
 
     public function testReturnNothingWhenAccessingUnknownParameter()
     {
-        $this->assertNull(Form::of([])->get('foo')->match(
+        $this->assertNull(Query::of([])->get('foo')->match(
             static fn($foo) => $foo,
             static fn() => null,
         ));
@@ -43,19 +43,19 @@ class FormTest extends TestCase
     {
         $this->assertSame(
             [1, 2, 3],
-            Form::of(['foo' => [1, 2, 3]])->list('foo')->match(
+            Query::of(['foo' => [1, 2, 3]])->list('foo')->match(
                 static fn($list) => $list->data(),
                 static fn() => null,
             ),
         );
         $this->assertNull(
-            Form::of(['foo' => 'bar'])->list('foo')->match(
+            Query::of(['foo' => 'bar'])->list('foo')->match(
                 static fn($list) => $list,
                 static fn() => null,
             ),
         );
         $this->assertNull(
-            Form::of(['foo' => [0 => 1, 2 => 3]])->list('foo')->match(
+            Query::of(['foo' => [0 => 1, 2 => 3]])->list('foo')->match(
                 static fn($list) => $list,
                 static fn() => null,
             ),
@@ -65,20 +65,20 @@ class FormTest extends TestCase
     public function testDictionary()
     {
         $this->assertNull(
-            Form::of(['foo' => [1, 2, 3]])->dictionary('foo')->match(
+            Query::of(['foo' => [1, 2, 3]])->dictionary('foo')->match(
                 static fn($list) => $list,
                 static fn() => null,
             ),
         );
         $this->assertNull(
-            Form::of(['foo' => 'bar'])->dictionary('foo')->match(
+            Query::of(['foo' => 'bar'])->dictionary('foo')->match(
                 static fn($list) => $list,
                 static fn() => null,
             ),
         );
         $this->assertSame(
             [0 => 1, 2 => 3],
-            Form::of(['foo' => [0 => 1, 2 => 3]])->dictionary('foo')->match(
+            Query::of(['foo' => [0 => 1, 2 => 3]])->dictionary('foo')->match(
                 static fn($list) => $list->data(),
                 static fn() => null,
             ),
@@ -87,7 +87,7 @@ class FormTest extends TestCase
 
     public function testNestedGet()
     {
-        $form = Form::of([
+        $query = Query::of([
             'foo' => [
                 'bar' => [
                     'baz' => '42',
@@ -97,7 +97,7 @@ class FormTest extends TestCase
 
         $this->assertSame(
             '42',
-            $form
+            $query
                 ->dictionary('foo')
                 ->flatMap(static fn($foo) => $foo->dictionary('bar'))
                 ->flatMap(static fn($bar) => $bar->get('baz'))